sandbox/src/test/java/ru/stqa/pft/sandbox/PrimesTest.java
package ru.stqa.pft.sandbox;

import org.testng.Assert;
import org.testng.annotations.Test;

public class PrimesTest {

  @Test
  public void testPrime() {
    Assert.assertTrue(Primes.isPrimeFast(Integer.MAX_VALUE)); // Проверяем максимально возможное число на простоту (Оно простое)
  }

  @Test
  public void testNonPrime() {
    Assert.assertFalse(Primes.isPrime(Integer.MAX_VALUE - 2)); //Проверка НЕпростого числа
  }

  @Test (enabled = false)
  public void testPrimeLong() {
    long n = Integer.MAX_VALUE;
    Assert.assertTrue(Primes.isPrime(n)); // Проверяем максимально возможное число на простоту (Оно простое)
  }


}
